% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cband.R
\name{cband}
\alias{cband}
\title{Create a simultaneous confidence band}
\usage{
cband(
  formula,
  alpha = 0.05,
  predgen = pgen_gam,
  newdata,
  data,
  method = "Empirical bootstrap",
  B = 500,
  robust_var = FALSE,
  weighted_boot = FALSE,
  ...
)
}
\arguments{
\item{formula}{The formula giving the specification.}

\item{alpha}{The significance level of the confidence band.}

\item{predgen}{The predictor generator. Can be the function or
character name of the function. See ?predictor_generators for more
information.}

\item{newdata}{The grid at which to give predictions.}

\item{data}{The data.}

\item{method}{The confidence band method.}

\item{B}{Number of bootstrap replications.}

\item{robust_var}{if TRUE, use robust variance estimation.}

\item{weighted_boot}{If TRUE, a multiplier bootstrap is used with normalized
i.i.d. standard exponential multipliers. If FALSE, a non-parametric
bootstrap is used. See Kosorok (2008) page 20.}

\item{...}{dots are passed to predgen.}
}
\value{
A data frame with columns "lower", "est", and "upper".
}
\description{
Create a simultaneous confidence band
}
\examples{
x <- rnorm(1000)
y <- rnorm(1000)
dframe <- data.frame(x = x, y = y)
# choose small 'B' to make quick; in practice B should be much larger.
cband_ex1 <- cband(formula = y ~ s(x), data = dframe, B = 10)
# for pgen_loc, we must use the verbose locfit::lp(x) and not just lp(x)
cband_ex2 <- cband(formula = y ~ locfit::lp(x), data = dframe,
                   predgen = pgen_loc, B = 10)

}
