% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgens.R
\name{pgen_gam}
\alias{pgen_gam}
\title{Estimation using mgcv::gam}
\usage{
pgen_gam(data_orig, formula, ...)
}
\arguments{
\item{data_orig}{The user's original data passed to the client function.}

\item{formula}{The user's original formula passed to the client function.}

\item{...}{dots are passed to gam.}
}
\description{
Estimation using mgcv::gam
}
