% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgens.R
\name{pgen_gam_deriv}
\alias{pgen_gam_deriv}
\title{Estimation of derivative using mgcv::gam}
\usage{
pgen_gam_deriv(data_orig, formula, deriv_order, ...)
}
\arguments{
\item{data_orig}{The user's original data passed to the client function.}

\item{formula}{The user's original formula passed to the client function.}

\item{deriv_order}{order of derivative.}

\item{...}{dots are passed to gam.}
}
\description{
Estimation of derivative using mgcv::gam
}
