% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgens.R
\name{pgen_glm}
\alias{pgen_glm}
\title{Estimation using stats::glm}
\usage{
pgen_glm(data_orig, formula, ...)
}
\arguments{
\item{data_orig}{The user's original data passed to the client function.}

\item{formula}{The user's original formula passed to the client function.}

\item{...}{dots are passed to glm.}
}
\description{
Estimation using stats::glm
}
