% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgens.R
\name{pgen_kde}
\alias{pgen_kde}
\title{Estimation using np::npudens}
\usage{
pgen_kde(data_orig, formula, ...)
}
\arguments{
\item{data_orig}{The user's original data passed to the client function.
The data cannot contain columns of variables not in the formula.}

\item{formula}{The user's original formula passed to the client function.
Must be of the form '~x'.}

\item{...}{dots are passed to npudens.}
}
\description{
Estimation using np::npudens
}
