% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgens.R
\name{pgen_lm}
\alias{pgen_lm}
\title{Estimation using stats::lm}
\usage{
pgen_lm(data_orig, formula, ...)
}
\arguments{
\item{data_orig}{The user's original data passed to the client function.}

\item{formula}{The user's original formula passed to the client function.}

\item{...}{dots are passed to lm.}
}
\description{
Estimation using stats::lm
}
