% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgens.R
\name{pgen_loc}
\alias{pgen_loc}
\title{Estimation using locfit::locfit}
\usage{
pgen_loc(data_orig, formula, deriv_order = 0, ...)
}
\arguments{
\item{data_orig}{The user's original data passed to the client function.}

\item{formula}{The user's original formula passed to the client function.
The formula must use the verbose 'locfit::lp(x)', or the user must
run "library(locfit)" before using this pgen.}

\item{deriv_order}{order of derivative.}

\item{...}{dots are passed to locfit.}
}
\description{
Estimation using locfit::locfit
}
