% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgens.R
\name{pgen_loess}
\alias{pgen_loess}
\title{Estimation using stats::loess}
\usage{
pgen_loess(data_orig, formula, ...)
}
\arguments{
\item{data_orig}{The user's original data passed to the client function.}

\item{formula}{The user's original formula passed to the client function.}

\item{...}{dots are passed to loess.}
}
\description{
Estimation using stats::loess
}
