% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgens.R
\name{pgen_npreg}
\alias{pgen_npreg}
\title{Estimation using np::npregbw}
\usage{
pgen_npreg(data_orig, formula, regtype_ = "ll", ...)
}
\arguments{
\item{data_orig}{The user's original data passed to the client function.}

\item{formula}{The user's original formula passed to the client function.}

\item{regtype_}{use local linear ("ll") or local constant, i.e., Nadaraya-Watson.}

\item{...}{dots are passed to npregbw.}
}
\description{
Use npreg's local linear (regtype = "ll") estimation.
}
