% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgens.R
\name{pgen_rq}
\alias{pgen_rq}
\title{Estimation quantreg::rq}
\usage{
pgen_rq(data_orig, formula, ...)
}
\arguments{
\item{data_orig}{The user's original data passed to the client function.}

\item{formula}{The user's original formula passed to the client function.}

\item{...}{dots are passed to rq.}
}
\description{
Estimation quantreg::rq
}
